+++ doc/source/changelog.rst
@@ -1,7 +1,13 @@
 Changelog
 =========
 
-2.2.0 - Unreleased
+2.3.0 - Unreleased
+------------------
+- ``File.__getitem__`` now returns bytearray instead of bytes. This prevents a memcpy
+  when deserializing numpy arrays with dask. (:pr:`TODO`) `Guido Imperiale`_
+
+
+2.2.0 - 2022-04-28
 ------------------
 - Added ``python_requires`` to ``setup.py`` (:pr:`60`) `Carlos Cordoba`_
 - Added type annotations (:pr:`62`) `Guido Imperiale`_